move out ticktype
