+++ gameserver/src/main/java/com/lhf/game/EffectPersistence.java
@@ -6,23 +6,6 @@ import java.util.StringJoiner;
 import com.lhf.server.interfaces.NotNull;
 
 public class EffectPersistence implements Comparable<EffectPersistence> {
-    public enum TickType {
-        INSTANT, ACTION, BATTLE, ROOM, CONDITIONAL;
-
-        public static TickType getTickType(String value) {
-            for (TickType type : values()) {
-                if (type.toString().equalsIgnoreCase(value)) {
-                    return type;
-                }
-            }
-            return null;
-        }
-
-        public static boolean isTickType(String value) {
-            return TickType.getTickType(value) != null;
-        }
-    }
-
     public class Ticker {
         private int countdown;
