+++ gameserver/src/main/java/com/lhf/game/TickType.java
@@ -1,0 +1,18 @@
+package com.lhf.game;
+
+public enum TickType {
+    INSTANT, ACTION, BATTLE, ROOM, CONDITIONAL;
+
+    public static TickType getTickType(String value) {
+        for (TickType type : values()) {
+            if (type.toString().equalsIgnoreCase(value)) {
+                return type;
+            }
+        }
+        return null;
+    }
+
+    public static boolean isTickType(String value) {
+        return TickType.getTickType(value) != null;
+    }
+}
