+++ gameserver/src/main/java/com/lhf/game/item/concrete/equipment/RustyDagger.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Set;
 
 import com.lhf.game.EffectPersistence;
-import com.lhf.game.EffectPersistence.TickType;
 import com.lhf.game.EffectResistance;
+import com.lhf.game.TickType;
 import com.lhf.game.creature.CreatureEffectSource;
 import com.lhf.game.dice.DamageDice;
 import com.lhf.game.dice.DieType;
