+++ gameserver/src/main/java/com/lhf/game/item/concrete/equipment/Shield.java
@@ -3,7 +3,7 @@ package com.lhf.game.item.concrete.equipment;
 import java.util.Collections;
 
 import com.lhf.game.EffectPersistence;
-import com.lhf.game.EffectPersistence.TickType;
+import com.lhf.game.TickType;
 import com.lhf.game.creature.CreatureEffectSource;
 import com.lhf.game.enums.EquipmentSlots;
 import com.lhf.game.enums.EquipmentTypes;
