+++ gameserver/src/main/java/com/lhf/game/lewd/AfterGlow.java
@@ -1,7 +1,7 @@
 package com.lhf.game.lewd;
 
 import com.lhf.game.EffectPersistence;
-import com.lhf.game.EffectPersistence.TickType;
+import com.lhf.game.TickType;
 import com.lhf.game.creature.Creature;
 import com.lhf.game.creature.CreatureEffect;
 import com.lhf.game.creature.CreatureEffectSource;
