+++ gameserver/src/main/java/com/lhf/game/magic/concrete/Thaumaturgy.java
@@ -6,9 +6,9 @@ import java.util.Set;
 
 import com.lhf.Taggable;
 import com.lhf.game.EffectPersistence;
-import com.lhf.game.EffectPersistence.TickType;
 import com.lhf.game.EffectResistance;
 import com.lhf.game.EffectResistance.TargetResistAmount;
+import com.lhf.game.TickType;
 import com.lhf.game.creature.Creature;
 import com.lhf.game.creature.vocation.Vocation.VocationName;
 import com.lhf.game.enums.Attributes;
